requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic
qrcode[pil]
python-multipart
python-dotenv